tidy up android
